+++ pkg/controller/nativeResourceWorker.go
@@ -377,22 +377,9 @@ func (ctlr *Controller) getServicePort(
 		return fmt.Errorf("Informer not found for namespace: %v", route.Namespace), port
 	}
 	svcIndexer := nrInf.svcInformer.GetIndexer()
-	svcName := route.Spec.To.Name
-	if route.Spec.Port != nil {
-		strVal := route.Spec.Port.TargetPort.StrVal
-		if strVal == "" {
-			port = route.Spec.Port.TargetPort.IntVal
-		} else {
-			port, err = ctlr.getResourceServicePort(route.Namespace, svcName, svcIndexer, strVal, resource.ResourceTypeRoute)
-			if nil != err {
-				return fmt.Errorf("Error while processing port for route %s: %v", route.Name, err), port
-			}
-		}
-	} else {
-		port, err = ctlr.getResourceServicePort(route.Namespace, svcName, svcIndexer, "", resource.ResourceTypeRoute)
-		if nil != err {
-			return fmt.Errorf("Error while processing port for route %s: %v", route.Name, err), port
-		}
+	port, err = ctlr.getResourceServicePortForRoute(svcIndexer, route)
+	if nil != err {
+		return fmt.Errorf("Error while processing port for route %s: %v", route.Name, err), port
 	}
 	log.Debugf("Port %v found for route %s", port, route.Name)
 	return nil, port